feat: Implement useLinks hook
