+++ .github/workflows/build.yml
@@ -41,6 +41,7 @@ jobs:
         run: npm run build
         env:
           VITE_APP_PASSWORD: ${{ secrets.VITE_APP_PASSWORD }}
+          VITE_FIREBASE_DB_URL: ${{ secrets.VITE_FIREBASE_DB_URL }}
       - name: Setup Pages
         uses: actions/configure-pages@v3
       - name: Upload artifact
